Add a text view to the content view.
The original content view has a text label at the top with the name of the file and a Done button at the bottom. I'm keeping them because they don't mess anything up and I need something like the Done button to go back to the document browser when I'm done typing.